requirements.txt
# requirements.txt
flask
flask-cors
python-dotenv
ultralytics
opencv-python-headless
numpy